--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,12 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyUnresolvedReferencesInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredIdentifiers">
-        <list>
-          <option value="picc_ui_rc" />
-        </list>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
