--- Enum/LanguageEnum.py
@@ -1,14 +1,0 @@
-# 定义一个枚举类
-from enum import Enum
-
-class Languages(Enum):
-    Java = 'Java'
-    Python = 'Python'
-    C_plus = 'C++'
-    C = 'C'
-    Go = 'Go'
-    JavaScript = 'JavaScript'
-    PHP = 'PHP'
-    Ruby = 'Ruby'
-    shell = 'shell'
-    Typescript = 'Typescript'
