+++ crawler/entity/project.py
@@ -20,25 +20,29 @@ class Project:
     GitHub 仓库项目实体类
     """
 
-    def __init__(self, item_name: str, stars: int, login_name: str, repository: str, description: str):
+    def __init__(self, item_name: str, stars: int, login_name: str, repository: str, description: str,size: int,id: str,full_name: str,default_branch: str):
         self.item_name = item_name  # 项目名称
         self.stars = stars  # 收藏数量
         self.login_name = login_name  # 登录名
         self.repository = repository  # 项目仓库网址
         self.description = description  # 项目仓库描述
+        self.size = size  # 项目大小
+        self.id = id  # 项目ID
+        self.full_name = full_name  # 项目所在用户名
+        self.default_branch = default_branch  # 项目的默认分支
 
     def to_list(self):
         """
         将实例的属性转换为列表
         """
-        return (self.item_name, self.stars, self.login_name, self.repository, self.description)
+        return (self.item_name, self.stars, self.login_name, self.repository, self.description,self.size,self.id,self.full_name,self.default_branch)
 
     @staticmethod
     def header():
         """
         返回 CSV 文件的表头
         """
-        return ['ItemName', 'Stars', 'LoginName', 'Repository', 'Description']
+        return ['ItemName', 'Stars', 'LoginName', 'Repository', 'Description','Size','Id','FullName','DefaultBranch']
 
     def __str__(self):
-        return 'Project: {' + f'item_name: {self.item_name}, start: {self.stars}, login_name: {self.login_name}, repository: {self.repository}, description: {self.description}' + '}'
+        return 'Project: {' + f'item_name: {self.item_name}, start: {self.stars}, login_name: {self.login_name}, repository: {self.repository}, description: {self.description}, size: {self.size}, id: {self.id}, full_name: {self.full_name},default_branch: {self.default_branch}' + '}'
