--- gift-main/.github/workflows/license.yml
@@ -1,14 +1,0 @@
-name: License Check
-
-on: [ push, pull_request, workflow_dispatch ]
-
-permissions:
-  contents: read
-
-jobs:
-  license:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Check License
-        uses: apache/skywalking-eyes@main
