--- gift-main/.gitignore
@@ -1,5 +1,0 @@
-.idea/
-target/
-**/config.yaml
-**/mybatis-config.xml
-**/**.pyc
