--- gift-main/README.md
@@ -1,2 +1,0 @@
-# gift
-gift - Github 开源项目搜索平台
