--- gift-main/spark/src/main/resources/hdfs-site.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
-<!--
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License. See accompanying LICENSE file.
--->
-
-<!-- Put site-specific property overrides in this file. -->
-
-<configuration>
-    <property>
-        <name>dfs.datanode.data.dir.perm</name>
-        <value>700</value>
-    </property>
-    <property>
-        <name>dfs.namenode.name.dir</name>
-        <value>/data/nn</value>
-    </property>
-    <property>
-        <name>dfs.namenode.hosts</name>
-        <value>master,slave1,slave2</value>
-    </property>
-    <property>
-        <name>dfs.client.use.datanode.hostname</name>
-        <value>true</value>
-    </property>
-    <property>
-        <name>dfs.blocksize</name>
-        <value>268435456</value>
-    </property>
-    <property>
-        <name>dfs.namenode.handler.count</name>
-        <value>100</value>
-    </property>
-    <property>
-        <name>dfs.datanode.data.dir</name>
-        <value>/data/dn</value>
-    </property>
-    <property>
-        <name>dfs.permissions</name>
-        <value>true</value>
-    </property>
-    <property>
-        <name>dfs.datanode.hostname</name>
-        <value>master</value>
-    </property>
-</configuration>
