--- gift-main/spark/src/main/resources/mybatis-config.bak.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE configuration
-        PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-config.dtd">
-
-<configuration>
-    <environments default="development">
-        <environment id="development">
-            <transactionManager type="JDBC"/>
-            <dataSource type="POOLED">
-                <property name="driver" value="com.mysql.cj.jdbc.Driver"/>
-                <property name="url" value="jdbc:mysql://localhost:3306/test?useSSL=false"/>
-                <property name="username" value="root"/>
-                <property name="password" value=""/>
-            </dataSource>
-        </environment>
-    </environments>
-    <mappers>
-        <mapper resource="org/giftorg/spark.mapper/ProjectMapper.xml"/>
-    </mappers>
-</configuration>
