--- gift-main/spark/src/main/resources/org/giftorg/spark.mapper/ProjectMapper.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="org.giftorg.spark.mapper.ProjectMapper">
-    <select id="selectOne" resultType="org.giftorg.spark.entity.Project">
-        select * from projects where id = #{id};
-    </select>
-    <select id="selectList" resultType="org.giftorg.spark.entity.Project">
-        select * from projects;
-    </select>
-</mapper>
