--- gift-main/spark/src/test/java/org/giftorg/spark/utils/GitUtilTest.java
@@ -1,44 +1,0 @@
-/**
- * Copyright 2023 GiftOrg Authors
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.giftorg.spark.utils;
-
-import cn.hutool.core.io.FileUtil;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.Test;
-
-import java.io.IOException;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@Slf4j
-class GitUtilTest {
-    @Test
-    void testGitClone() throws IOException {
-        String remote = "https://github.com/crazycodeboy/TakePhoto";
-        String local = FileUtil.getAbsolutePath(PathUtil.join("git_test", PathUtil.base(remote)));
-
-        Boolean ok = GitUtil.gitClone(remote, local);
-        assertEquals(true, ok);
-        assertTrue(FileUtil.exist(local));
-
-        ok = FileUtil.del(local);
-        assertEquals(true, ok);
-    }
-}
