--- gift-main/spark/src/test/java/org/giftorg/spark/utils/PathUtilTest.java
@@ -1,47 +1,0 @@
-/**
- * Copyright 2023 GiftOrg Authors
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.giftorg.spark.utils;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class PathUtilTest {
-    @Test
-    void testJoin() {
-        String path = PathUtil.join("a", "b\\", "c/", "d");
-
-        String expected = "a/b/c/d";
-        if (isWindows()) expected = "a\\b\\c\\d";
-
-        assertEquals(expected, path);
-    }
-
-    @Test
-    void testBase() {
-        String path = PathUtil.base("https://site.com/a");
-        assertEquals("a", path);
-    }
-
-    public static Boolean isWindows() {
-        String os = System.getProperty("os.name").toLowerCase();
-        return os.contains("win");
-    }
-}
